move included/excluded files initialization to CSVFileManager
