+++ src/main/java/org/mitre/synthea/export/CSVExporter.java
@@ -14,9 +14,7 @@ import java.math.BigDecimal;
 import java.math.RoundingMode;
 import java.nio.file.Path;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Calendar;
-import java.util.Collections;
 import java.util.Comparator;
 import java.util.GregorianCalendar;
 import java.util.Iterator;
@@ -90,61 +88,9 @@ public class CSVExporter {
   }
 
   void init() {
-    String includedFilesStr = Config.get("exporter.csv.included_files", "").trim();
-    String excludedFilesStr = Config.get("exporter.csv.excluded_files", "").trim();
-
-    List<String> includedFiles = Collections.emptyList();
-    List<String> excludedFiles = Collections.emptyList();
-
-    if (!includedFilesStr.isEmpty() && !excludedFilesStr.isEmpty()) {
-      includedFiles = propStringToList(includedFilesStr);
-      excludedFiles = propStringToList(excludedFilesStr);
-
-      // Check if there is any overlap
-      for (String includedFile : includedFiles) {
-        if (excludedFiles.contains(includedFile)) {
-          System.err.println("ERROR! CSV exporter is set to include and exclude the same file: "
-                  + includedFile);
-          throw new IllegalArgumentException(
-                  "CSV exporter cannot include and exclude the same file: " + includedFile);
-        }
-      }
-    } else {
-      if (!includedFilesStr.isEmpty()) {
-        includedFiles = propStringToList(includedFilesStr);
-
-        if (!includedFiles.contains("patients.csv")) {
-          System.err.println("WARNING! CSV exporter is set to not include patients.csv!");
-          System.err.println("This is probably not what you want!");
-        }
-
-      } else {
-        excludedFiles = propStringToList(excludedFilesStr);
-      }
-    }
-
-    fileManager = new CSVFileManager(includedFiles, excludedFiles);
+    fileManager = new CSVFileManager();
 
     this.transactionId = new AtomicLong();
-  }
-
-  /**
-   * Helper function to convert a list of files directly from synthea.properties to filenames.
-   * @param fileListString String directly from Config, ex "patients.csv,conditions , procedures"
-   * @return normalized list of filenames as strings
-   */
-  private static List<String> propStringToList(String fileListString) {
-    List<String> files = Arrays.asList(fileListString.split(","));
-    // normalize filenames -- trim, lowercase, add .csv if not included
-    files = files.stream().map(f -> {
-      f = f.trim().toLowerCase();
-      if (!f.endsWith(".csv")) {
-        f = f + ".csv";
-      }
-      return f;
-    }).collect(Collectors.toList());
-
-    return files;
   }
 
   /**
